+++ tests/formatter/input-output-pairs/comment_at_the_end_of_file_237.out.gd
@@ -1,0 +1,3 @@
+func foo():
+	var a = 1
+#	# var i = 2